backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.129.0",
    "grandalf>=0.8",
    "langchain-openai>=1.1.9",
    "langgraph>=1.0.8",
    "pip>=26.0.1",
    "presidio-analyzer>=2.2.361",
    "presidio-anonymizer>=2.2.361",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.27.1",
    "pymupdf4llm>=0.3.4",
    "python-dotenv>=1.2.1",
    "python-magic-bin>=0.4.14",
    "python-multipart>=0.0.22",
    "redis>=7.1.1",
    "spacy>=3.8.11",
    "tzdata>=2025.3",
    "uvicorn>=0.40.0",
]
